requirements.txt
django
mysqlclient
djangorestframework
django-cors-headers
drf-yasg
django-environ
celery
